backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "juapesa-backend"
version = "0.1.0"
description = "Jua Pesa — AI Liquidity Network backend (FastAPI)"
authors = [{ name = "Jua Pesa Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7",
  "pydantic-settings>=2.3.4",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "requests>=2.32.0",
  "redis>=5.0.7",
  "sqlalchemy>=2.0.30",
  "psycopg2-binary>=2.9.9",
  "prometheus-client>=0.20.0",
  "typing-extensions>=4.11.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=0.23.7",
  "ruff>=0.5.4",
  "mypy>=1.10.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --cov=app --cov-report=term-missing --cov-fail-under=100"
pythonpath = ["app"]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = false
